requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
pymilvus
llama-cpp-python
